requirements.txt
pandas
unidecode
PyPDF2
Pillow 
customtkinter